pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "newsight"
version = "1.0.0"
description = "RAG와 LLM을 활용한 뉴스/논문 요약 & 해설 서비스"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0", 
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.23",
    "aiosqlite>=0.19.0",
]

[project.optional-dependencies]
ai = [
    "openai>=1.3.7",
    "langchain>=0.0.340",
    "langchain-openai>=0.0.2", 
    "langchain-community>=0.0.1",
    "sentence-transformers>=2.2.2",
    "chromadb>=0.4.18",
    "faiss-cpu>=1.7.4",
    "numpy>=1.25.2",
]

scraping = [
    "aiohttp>=3.8.0",
    "feedparser>=6.0.10",
    "beautifulsoup4>=4.12.2",
    "requests>=2.31.0",
    "python-dateutil>=2.8.2",
]

database = [
    "psycopg2-binary>=2.9.9",
    "alembic>=1.12.1",
]

background = [
    "celery>=5.3.4",
    "redis>=5.0.1",
    "aiofiles>=23.2.1",
]

security = [
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.1.2",
]

data = [
    "pandas>=2.1.4",
    "pytz>=2023.3",
]

# 모든 선택적 의존성
all = [
    "newsight[ai,scraping,database,background,security,data]"
]

# 개발용 의존성
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]

[tool.setuptools.packages.find]
where = ["backend"]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
